refactor: move otel initialization to tracing.js
