+++ api/package.json
@@ -61,9 +61,9 @@
     "postdeploy:datamart": "DEBUG=knex:query npm run datamart:create && DEBUG=knex:query npm run datamart:migrate",
     "postdeploy:maddo": "node scripts/wait-for-api-deployment && npm run datamart:migrate",
     "scalingo-postbuild": "node scripts/generate-cron > cron.json && node scripts/generate-procfile",
-    "dev": "nodemon --env-file-if-exists=.env -r ./src/shared/infrastructure/otel/otel-setup.js index.js",
+    "dev": "nodemon --env-file-if-exists=.env -r ./tracing.js index.js",
     "dev:maddo": "MADDO=true nodemon --env-file-if-exists=.env index.maddo.js",
-    "start": "node --env-file-if-exists=.env -r ./src/shared/infrastructure/otel/otel-setup.js index.js",
+    "start": "node --env-file-if-exists=.env -r ./tracing.js index.js",
     "start:inspect": "node --env-file-if-exists=.env --inspect index.js",
     "start:maddo": "MADDO=true node --env-file-if-exists=.env index.maddo.js",
     "start:watch": "npm run dev",
